backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[argon2]
python-jose[cryptography]
python-multipart
websockets
requests
groq
python-dotenv
sendgrid
